chore: decomposition of homepage and remove comments
