+++ src/app/page.module.sass
@@ -13,243 +13,4 @@
         &:hover
             filter: brightness(0.9)
 
-.sectionBanner
-    height: 500px
-    display: flex
-    flex-direction: column
-    gap: 20px
-    text-align: center
-    align-items: center
-    padding-top: 80px
-    background-repeat: no-repeat
-    background-size: contain
-    background-position: center 2px
-    background-color: #191919
-    > span
-        text-transform: uppercase
-        letter-spacing: 0.2rem
-        color: var(--gray)
-    > h1
-        font-size: 2.5rem
-        text-transform: uppercase
-        letter-spacing: 0.15rem
-    > p
-        color: var(--gray)
-        max-width: 350px
-    > button
-        background-color: var(--orange)
-        margin-top: 20px
-        color: white
-    @media screen and (min-width: 768px)
-        > h1
-            font-size: 3.2rem
-    @media screen and ( min-width: 1200px )
-        align-items: flex-start
-        text-align: left
-        padding-left: 200px
-        height: 700px
-        padding-top: 120px
-        background-position: center
-        > h1
-            max-width: 500px
-        > p
-            font-size: 1.1rem
-            max-width: 400px
-
-.sectionCategories
-    display: flex
-    flex-direction: column
-    align-items: center
-    margin: 0 auto
-    margin-top: 80px
-    gap: 80px
-    max-width: 90%
-    .categoryCard
-        display: flex
-        flex-direction: column
-        align-items: center
-        justify-content: flex-end
-        padding-bottom: 20px
-        gap: 10px
-        width: 100%
-        height: 200px
-        background-color: var(--gray-bg)
-        border-radius: 10px
-        color: black
-        text-transform: uppercase
-        position: relative
-        &:hover
-            .imageContainer
-                top: -63px
-        > h2
-            font-size: 0.95rem
-        > button
-            display: flex
-            justify-content: center
-            align-items: center
-            font-size: 0.85rem
-            font-weight: 600
-            color: var(--gray)
-            > svg path
-                stroke: var(--orange)
-            &:hover
-                color: var(--orange)
-    .imageContainer
-        position: absolute
-        top: -60px
-        transition: all 0.3s ease
-        > img
-            width: 180px
-            height: auto
-    @media screen and (min-width: 768px)
-        flex-direction: row
-        justify-content: center
-        margin-top: 100px
-        width: 100%
-        gap: 5px
-        .categoryCard
-            height: 180px
-            > h2
-                font-size: 1rem
-            .imageContainer
-                top: -80px
-                > img
-                    width: 200px
-                    height: auto
-            &:hover
-                .imageContainer
-                    top: -83px
-                    height: auto
-
-.sectionFeatured
-    display: flex
-    flex-direction: column
-    align-items: center
-    margin: 0 auto
-    margin-top: 30px
-    width: 90%
-    .cardOne
-        display: flex
-        flex-direction: column
-        align-items: center
-        width: 100%
-        border-radius: 10px
-        background-color: var(--orange)
-        padding: 50px 20px
-        .textContainer
-            display: flex
-            flex-direction: column
-            align-items: center
-            text-align: center
-            gap: 20px
-            > h2
-                font-size: 2.2rem
-                width: 70%
-            > p
-                opacity: 0.8
-        button
-            background-color: black
-            color: white
-        .imageContainer
-            width: 280px
-            > img
-                width: 100%
-                height: auto
-        @media screen and (min-width: 768px)
-            margin: 0 40px
-            p
-                width: 50%
-        @media screen and (min-width: 1200px)
-            display: grid
-            grid-template-columns: 55% 1fr
-            overflow: hidden
-            padding: 100px 50px
-            height: 500px
-            .imageContainer
-                width: 700px
-                position: relative
-                > img
-                    position: absolute
-                    top: -180px
-                    right: 30px
-            .textContainer
-                align-items: flex-start
-                text-align: left
-                > h2
-                    font-size: 2.5rem
-                > p
-                    width: 350px
-
-    .cardTwo, .cardThree
-        color: black
-        text-transform: uppercase
-        button 
-            width: 150px
-            border: 2px black solid
-            &:hover
-                background-color: black
-                color: white
-
-    .cardTwo
-        display: flex
-        flex-direction: column
-        gap: 10px
-        width: 100%
-        height: 300px
-        margin-top: 30px
-        border-radius: 10px
-        background-size: cover
-        background-repeat: no-repeat
-        background-position: center
-        padding: 100px 0 0 40px
-
-        @media screen and (min-width: 768px)
-            padding-left: 80px
-            margin: 30px auto
-    .cardThree
-        display: flex
-        flex-direction: column
-        align-items: center
-        width: 100%
-        margin: 30px 0
-        .imageContainer, .textContainer
-            border-radius: 10px
-            width: 100%
-            height: 200px
-            overflow: hidden
-            img 
-                width: 100%
-                height: auto
-                border-radius: 10px
-        .textContainer
-            background-color: var(--gray-bg)
-            margin-top: 30px
-            display: flex
-            flex-direction: column
-            align-items: center
-            justify-content: center
-            button
-                margin-top: 10px
-        @media screen and (min-width: 768px)
-            display: grid
-            grid-template-columns: 1fr 1fr
-            gap: 20px
-            margin-top: 0
-            .imageContainer, .textContainer
-                height: 300px
-                margin-top: 0
-                img
-                    height: 100%
-                    width: auto
-                    object-position: -20px
-            .textContainer
-                align-items: flex-start
-                padding-left: 50px
-        @media screen and (min-width: 1120px)
-            .imageContainer, .textContainer
-                img
-                    height: auto
-                    width: 100%
-                    object-position: center
-
