+++ src/components/Homepage/subcomponents/Categories.module.sass
@@ -1,0 +1,64 @@
+.sectionCategories
+    display: flex
+    flex-direction: column
+    align-items: center
+    margin: 0 auto
+    margin-top: 80px
+    gap: 80px
+    max-width: 90%
+    .categoryCard
+        display: flex
+        flex-direction: column
+        align-items: center
+        justify-content: flex-end
+        padding-bottom: 20px
+        gap: 10px
+        width: 100%
+        height: 200px
+        background-color: var(--gray-bg)
+        border-radius: 10px
+        color: black
+        text-transform: uppercase
+        position: relative
+        &:hover
+            .imageContainer
+                top: -63px
+        > h2
+            font-size: 0.95rem
+        > button
+            display: flex
+            justify-content: center
+            align-items: center
+            font-size: 0.85rem
+            font-weight: 600
+            color: var(--gray)
+            > svg path
+                stroke: var(--orange)
+            &:hover
+                color: var(--orange)
+    .imageContainer
+        position: absolute
+        top: -60px
+        transition: all 0.3s ease
+        > img
+            width: 180px
+            height: auto
+    @media screen and (min-width: 768px)
+        flex-direction: row
+        justify-content: center
+        margin-top: 100px
+        width: 100%
+        gap: 5px
+        .categoryCard
+            height: 180px
+            > h2
+                font-size: 1rem
+            .imageContainer
+                top: -80px
+                > img
+                    width: 200px
+                    height: auto
+            &:hover
+                .imageContainer
+                    top: -83px
+                    height: auto
