+++ src/components/Homepage/subcomponents/FeaturedSection.module.sass
@@ -1,0 +1,130 @@
+.sectionFeatured
+    display: flex
+    flex-direction: column
+    align-items: center
+    margin: 0 auto
+    margin-top: 30px
+    width: 90%
+    .cardOne
+        display: flex
+        flex-direction: column
+        align-items: center
+        width: 100%
+        border-radius: 10px
+        background-color: var(--orange)
+        padding: 50px 20px
+        .textContainer
+            display: flex
+            flex-direction: column
+            align-items: center
+            text-align: center
+            gap: 20px
+            > h2
+                font-size: 2.2rem
+                width: 70%
+            > p
+                opacity: 0.8
+        button
+            background-color: black
+            color: white
+        .imageContainer
+            width: 280px
+            > img
+                width: 100%
+                height: auto
+        @media screen and (min-width: 768px)
+            margin: 0 40px
+            p
+                width: 50%
+        @media screen and (min-width: 1200px)
+            display: grid
+            grid-template-columns: 55% 1fr
+            overflow: hidden
+            padding: 100px 50px
+            height: 500px
+            .imageContainer
+                width: 700px
+                position: relative
+                > img
+                    position: absolute
+                    top: -180px
+                    right: 30px
+            .textContainer
+                align-items: flex-start
+                text-align: left
+                > h2
+                    font-size: 2.5rem
+                > p
+                    width: 350px
+
+    .cardTwo, .cardThree
+        color: black
+        text-transform: uppercase
+        button 
+            width: 150px
+            border: 2px black solid
+            &:hover
+                background-color: black
+                color: white
+
+    .cardTwo
+        display: flex
+        flex-direction: column
+        gap: 10px
+        width: 100%
+        height: 300px
+        margin-top: 30px
+        border-radius: 10px
+        background-size: cover
+        background-repeat: no-repeat
+        background-position: center
+        padding: 100px 0 0 40px
+
+        @media screen and (min-width: 768px)
+            padding-left: 80px
+            margin: 30px auto
+    .cardThree
+        display: flex
+        flex-direction: column
+        align-items: center
+        width: 100%
+        margin: 30px 0
+        .imageContainer, .textContainer
+            border-radius: 10px
+            width: 100%
+            height: 200px
+            overflow: hidden
+            img 
+                width: 100%
+                height: auto
+                border-radius: 10px
+        .textContainer
+            background-color: var(--gray-bg)
+            margin-top: 30px
+            display: flex
+            flex-direction: column
+            align-items: center
+            justify-content: center
+            button
+                margin-top: 10px
+        @media screen and (min-width: 768px)
+            display: grid
+            grid-template-columns: 1fr 1fr
+            gap: 20px
+            margin-top: 0
+            .imageContainer, .textContainer
+                height: 300px
+                margin-top: 0
+                img
+                    height: 100%
+                    width: auto
+                    object-position: -20px
+            .textContainer
+                align-items: flex-start
+                padding-left: 50px
+        @media screen and (min-width: 1120px)
+            .imageContainer, .textContainer
+                img
+                    height: auto
+                    width: 100%
+                    object-position: centers
