+++ src/lib/fetchGql.ts
@@ -1,11 +1,3 @@
-// const contentful = require('contentful')
-
-// export const client = contentful.createClient({
-//   space: process.env.CONTENTFUL_SPACE_ID,
-//   environment: 'master', // defaults to 'master' if not set
-//   accessToken: process.env.CONTENTFUL_ACCESS_TOKEN,
-// })
-
 export default async function fetchGql(query: string) {
   try {
     const response = await fetch(
